Preparing high-level dynamic call instruction (4): Changed pointers from int to own datatype, so it's being adjusted to 32/64-bit automatically, and we don't need it twice

+++ src/utils/CallStaticInit.kt
@@ -142,8 +142,8 @@ object CallStaticInit {
 
         // calculate how much new memory was used
         val timeI = System.nanoTime()
-        val allocationStart = engine.globals["allocationStart"]!!.toInt()
-        val allocationPointer = engine.globals["allocationPointer"]!!.toInt()
+        val allocationStart = engine.globals["allocationStart"]!!.toLong()
+        val allocationPointer = engine.globals["allocationPointer"]!!.toLong()
         LOGGER.info("Base Memory: ($allocationStart) ${allocationStart.formatFileSize()}")
         LOGGER.info("Allocated ${(allocationPointer - allocationStart).formatFileSize()} during StaticInit")
         LOGGER.info("Executed ${engine.instructionCounter} instructions for StaticInit")
